backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
bcrypt==4.1.2
python-dotenv==1.0.0
psycopg[binary]>=3.2.2
requests==2.31.0
twilio==8.10.0
firebase-admin==6.4.0
azure-cognitiveservices-speech==1.32.1
python-dateutil==2.8.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
reportlab==4.0.7
openpyxl==3.1.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
